requirements.txt
Flask
flask-cors
flask-pymongo
bcrypt
PyJWT
python-dotenv